modified code with kicki style of for loop using the (size_t) keyword

+++ c++_lab_1/calculateMMM.cpp
@@ -3,8 +3,8 @@ using namespace std;
 
 // sort array function
 int * sorted(int array[5]) {
-	for (int i=0; i<5;i++){
-		for(int j=i+1; j<5;j++){
+	for (size_t i=0; i<5;i++){
+		for(size_t j=i+1; j<5;j++){
 			if(array[i] > array[j]){
 				int temp = array[i];
 				array[i] = array[j];
@@ -20,9 +20,9 @@ int mode(int array[5]) {
 	int mode;
 
 // using for loop to count number occurrences and store in numCount array
-for(int i=0;i<5;i++){
+for(size_t i=0;i<5;i++){
 	int counter = 0;
-	for(int j=0;j<5;j++){
+	for(size_t j=0;j<5;j++){
 		if(array[i] == array[j]){
 			counter++;
 		}
@@ -30,18 +30,18 @@ for(int i=0;i<5;i++){
 	}
 } 
 
-// assigning numCount to tempCount
+// assigning maxOccurrence to the first value in numCount array
 int maxOccurrence = numCount[0];
 
-// loop through the tempCount to find the highest occurrence number
-for(int i=0;i<5;i++){
+// loop through the numCount to find the highest occurrence number
+for(size_t i=0;i<5;i++){
 	if(maxOccurrence < numCount[i]){
 		maxOccurrence = numCount[i];
 	}
 }
 
 // loop through numCount for most occurrence index and insert index in array
-for(int i=0;i<5;i++){
+for(size_t i=0;i<5;i++){
 	if(numCount[i] == maxOccurrence){
 		mode = array[i];
 		break;
@@ -54,7 +54,7 @@ return mode;
 // median function
 int median(int array[5]) {
 	int value;
-	for(int i=0;i<5;i++) {
+	for(size_t i=0;i<5;i++) {
 		if(i == 2){
 		 value=array[i];
 		}
@@ -68,7 +68,7 @@ int main() {
 	int sum = 0;
 
 // using for loop to get user input
-for(int i=0;i<5;i++){
+for(size_t i=0;i<5;i++){
 	cout<<"Enter num"<<i+1<<endl;
 	cin>>numsArray[i];
 }
